Azure/sensor/requirements.txt
PluginBase
requests
pyyaml